requirements.txt
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
azure-core==1.32.0
azure-ai-documentintelligence==1.0.1
langchain-core==0.3.45
langchain_openai==0.3.9